refactored CurrentMonthSummaryChart; cleaned up CurrentMonthView folder
