+++ src/components/shared/LabeledMoneyInput.jsx
